requirements.txt
beautifulsoup4~=4.11.2
selenium~=4.8.2
webdriver-manager
azure.cosmos
azure.identity
amazoncaptcha